Shorcuts for Preset Zoom Positions
Created dedicated buttons for preset zoom positions for conveinience. As the zoom in and out buttons are iterative and can be slow for large zoom operations, we have created dedicated buttons for zooming in and out to levels of x1, x10, x50 and x100. Which are supported by the Samsung Galaxy S23 Ultra that was used in this project.